+++ src/clj_kondo/impl/analyzer/clojure_data_xml.clj
@@ -16,17 +16,25 @@
                       (token-node (symbol (name sexpr)))])
           :else alias-node)))
 
+(defn encode-uri [uri]
+  (java.net.URLEncoder/encode uri "UTF-8"))
+
+(defn uri-symbol [uri]
+  (symbol (encode-uri (str "xmlns." uri))))
+
 (defn analyze-alias-uri [ctx expr]
   (let [children (next (:children expr))]
     (common/analyze-children ctx children)
     (loop [children children]
       (when children
         (let [alias-node (first children)
-              ns-str-node (second children)]
+              ns-str-node (second children)
+              ns-str (utils/sexpr ns-str-node)
+              uri (uri-symbol ns-str)]
           (common/analyze-expression** ctx
                                        (list-node
                                         [(token-node 'alias)
                                          (->alias-node alias-node)
                                          (list-node [(token-node 'quote)
-                                                     (token-node (symbol (str (utils/sexpr ns-str-node))))])]))
+                                                     (token-node uri)])]))
           (recur (nnext children)))))))